Refactor utils to prepare for get_class.
