+++ oscarapi/serializers/login.py
@@ -1,7 +1,7 @@
 from django.contrib.auth import get_user_model, authenticate
 from rest_framework import serializers
 
-from oscarapi.utils import overridable
+from oscarapi.utils.settings import overridable
 
 
 User = get_user_model()
